Modernize codebase to C17 with zero-warning build (GCC & Clang)
- config.mak.linux: target C17 (-std=c17), add comprehensive warning flags
  (-Wall -Wextra -Wformat=2 -Wformat-security -Wstrict-overflow=2 -Wcast-align
   -Wpointer-arith -Wmissing-prototypes -Wshadow -Wno-cast-function-type),
  add -D_POSIX_C_SOURCE=200809L to COPTS for POSIX symbol visibility
- include/ripriv.h: map RtBoolean→int16_t, RtInt→int32_t via <stdint.h>
- include/ribhash.h: map RIB_INT32→int32_t, RIB_UINT32→uint32_t via <stdint.h>
- include/ribattrstore.h: fix invalid typedef syntax (parentheses around type)
- include/ribrdr.h: fix RibSetRITable/RibSetHintTable macros (NULL→0,
  uintptr_t cast to avoid -Waddress on global array pointer test)
- readrib/streamio.c: use S_ISDIR() instead of S_IFDIR bitmask; fix gzFile
  double-pointer (gzFile is already a pointer in modern zlib); fix use-after-free
  (free member caches before freeing rib struct); use kRIB_ERRRC_INT in
  int-returning functions
- readrib/error.c: fix bounds check logic (&&→||, sizeof(severity)→sizeof(severities));
  use %zu for size_t fields linecount/byteoffset
- readrib/readfunc.c: fix sign-compare in sizeof loops (cast to int/unsigned);
  fix kRIB_ERRRC_PTR→kRIB_ERRRC_INT in int-returning functions; add
  /* falls through */ annotation for intentional switch fallthrough
- readrib/paraml.c: rename inner 'c' variable to 'namecopy' to avoid -Wshadow
- readrib/declare.c: (no change needed after lowering -Wstrict-overflow to 2)
- readrib/errtest.c: suppress unused argc/argv with (void) casts
- readrib/vcache.c: fix sign-compare (unsigned vectorcache fields vs signed RtInt)
- readrib/strmtest.c: fix sign-compare in sizeof loops; use %zu for size_t buffer lengths
- ribhash/hash.c: fix sign-compare (cast type/code to match unsigned hash fields);
  fix sizeof loop; add -Wfree-nonheap-object pragma for checked non-stdlib free
- ribhash/hashtest.c: suppress unused argc/argv with (void) casts
- ributil/attrstore.c: fix use-after-free (save ->prev before _RibFree)
- ributil/trimstore.c: fix use-after-free (save ->prev before _RibFree)
- ributil/ribfixer.c: rename mmErrorHandler 'rib' param to 'hrib' (-Wshadow)
- ributil/ribfunc.c: fix sign-compare (cast RtInt to unsigned for sizes[] compare)
- sribw/write.c: fix int-to-pointer casts via uintptr_t intermediary
- pcutil/untab.c: use explicit format string for -Wformat-security
- pixNtiff/wtiff.c: suppress -Wmissing-braces for intentional IFD struct packing;
  change loop counter to unsigned long to match x/y dimensions

+++ readrib/strmtest.c
@@ -182,7 +182,7 @@ int main( void )
 
    RibSaveToBuffer( rib, RI_FALSE );
    if ((u = RibGetBufferLength(rib)) != 0)
-      printf("ERROR:  Buffer Length!=0 (buflen==%u)\n",u );
+      printf("ERROR:  Buffer Length!=0 (buflen==%zu)\n",u );
 #ifdef RIB_MEMORY_TESTS
    ss = (char*)_RibMalloc( 1234 + 1 );
 #else
@@ -199,7 +199,7 @@ int main( void )
          ss[i] = i%26 + 'A';
       }
       if ((u = RibGetBufferLength(rib)) != 1234)
-        printf("ERROR:  Buffer Length!=1234 (buflen==%u)\n",u );
+        printf("ERROR:  Buffer Length!=1234 (buflen==%zu)\n",u );
       else
       {
          s = RibCreateStringFromBuffer( rib, 0 );
@@ -241,7 +241,7 @@ int main( void )
 
       for (i=0; i<1234; i++)
       {
-         for (c=0; c<sizeof(RtInt); c++ )
+         for (c=0; c<(int)sizeof(RtInt); c++ )
          {
             RibPutChar( rib, ((char*)&I)[c] );
             ss[i*sizeof(RtInt)+c] = ((char*)&I)[c];
@@ -250,7 +250,7 @@ int main( void )
       u = RibGetBufferLength(rib);
       if ( u != 1234*sizeof(RtInt) )
       {
-         printf( "ERROR:  Buffer Length!=1234*sizeof(RtInt) (buflen==%u)\n",
+         printf( "ERROR:  Buffer Length!=1234*sizeof(RtInt) (buflen==%zu)\n",
                  u );
       }
       else
@@ -292,7 +292,7 @@ int main( void )
 
       for (i=0; i<1234; i++)
       {
-         for (c=0; c<sizeof(RtFloat); c++ )
+         for (c=0; c<(int)sizeof(RtFloat); c++ )
          {
             RibPutChar( rib, ((char*)&F)[c] );
             ss[i*sizeof(RtFloat)+c] = ((char*)&F)[c];
@@ -302,7 +302,7 @@ int main( void )
       if ( u != 1234*sizeof(RtFloat))
       {
          printf("ERROR:  " \
-                "Buffer Length!=1234*sizeof(RtFloat) (buflen==%u)\n",u );
+                "Buffer Length!=1234*sizeof(RtFloat) (buflen==%zu)\n",u );
       }
       else
       {
